Move new tests to correct framework
