+++ src/System.Collections/tests/Generic/Dictionary/Dictionary.Generic.Tests.netcoreapp.cs
@@ -418,5 +418,22 @@ namespace System.Collections.Tests
         }
 
         #endregion
+
+        #region ValueRef 
+
+        [Theory]
+        [MemberData(nameof(ValidCollectionSizes))]
+        public void Dictionary_Generic_ValueRef_ContainsAllCorrectValues(int count)
+        {
+            Dictionary<TKey, TValue> dictionary = (Dictionary<TKey, TValue>)GenericIDictionaryFactory(count);
+            IEnumerable<TKey> keys = ((IReadOnlyDictionary<TKey, TValue>)dictionary).Keys;
+
+            foreach (TKey key in keys)
+            {
+                Assert.Equal(dictionary[key], dictionary.ValueRef(key));
+            }
+        }
+
+        #endregion
     }
 }
